requirements.txt
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-bigquery==3.24.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
idna==3.7
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
six==1.16.0
urllib3==2.2.2
